Renamed file and refactored interface
